Fix unit tests after moving code to EE folder
